+++ packages/md-enhance/src/client/components/Mermaid.ts
@@ -1,3 +1,4 @@
+import { useMutationObserver } from "@vueuse/core";
 import { type MermaidConfig } from "mermaid";
 import {
   type VNode,
@@ -5,7 +6,6 @@ import {
   defineComponent,
   h,
   onMounted,
-  onUnmounted,
   ref,
   watch,
 } from "vue";
@@ -131,7 +131,7 @@ export default defineComponent({
       });
 
     onMounted(() => {
-      const html = document.querySelector("html")!;
+      const html = document.documentElement;
 
       const getDarkmodeStatus = (): boolean =>
         html.classList.contains("dark") ||
@@ -143,20 +143,18 @@ export default defineComponent({
       void renderMermaid();
 
       // watch darkmode change
-      const observer = new MutationObserver(() => {
-        isDarkmode.value = getDarkmodeStatus();
-      });
-
-      observer.observe(html, {
-        attributeFilter: ["class", "data-theme"],
-        attributes: true,
-      });
+      useMutationObserver(
+        html,
+        () => {
+          isDarkmode.value = getDarkmodeStatus();
+        },
+        {
+          attributeFilter: ["class", "data-theme"],
+          attributes: true,
+        }
+      );
 
       watch(isDarkmode, () => renderMermaid());
-
-      onUnmounted(() => {
-        observer.disconnect();
-      });
     });
 
     return (): VNode =>