+++ packages/rtl/rollup.config.ts
@@ -1,6 +1,3 @@
-import { rollupTypescript } from "../../scripts/rollup.js";
+import { bundle } from "../../scripts/rollup.js";
 
-export default [
-  ...rollupTypescript("node/index"),
-  ...rollupTypescript("client/config"),
-];
+export default [...bundle("node/index"), ...bundle("client/config")];